Fix au_write() to set 'errno' in the event the arguments are invalid. Check to see if the call to au_write() would cause more data to be written to the audit record than is available, including checking to make sure we don't use room reserved for the header and trailer.
Break out the logic in au_close()that flattens out the audit record
tokens into a buffer, as well as the logic to prepend a the audit
header and append the audit trailer.  The new function,
au_assemble(), places the contents of the tokens, ready to commit, in   
the buffer attached to the record.  It does not free the tokens,
which will be done by the caller.

Break out the logic used to free all memory associated with freeing  
an audit record, including all tokens, into au_teardown().

Implement au_close() in terms of these calls.  Comment on some
failure cases we need to look at carefully and that can maybe now be    
eliminated by virtue of space reservations for the header and trailer   
by au_write().

Implement au_close_buffer(), which writes the record to a memory
buffer rather than the system audit trail, which is useful for
writing test cases, as well as directly maintaining audit trails in    
user space.

[git-p4: depot-paths = "//depot/projects/trustedbsd/openbsm/": change = 75889]