When decoding the following token types, assume that the consumer wants network byte order IP addresses and port numbers, and in some cases other multi-byte IP header fields:
- AUT_IN_ADDR
- AUT_IP
- AUT_IPORT
- AUT_SOCKINET32
- AUT_SOCKET
- AUT_SOCKET32_EX

When printing these fields, convert from network byte order.

When encoding the following token types, assume the consumer
provides IP addresses and port numbers in network byte order:

- AUT_IN_ADDR
- AUT_IN_ADDR_EX
- AUT_SOCKINET32
- AUT_SOCKINET128

In general, prefer to hard-code the size of fields in terms of
the sizes they are defined in in the BSM spec.  It's not clear
whether this is really better or worse than defining it in
terms of local data types, since we will break nastily either
way.

[git-p4: depot-paths = "//depot/projects/trustedbsd/openbsm/": change = 92809]