Improve debugging output for auditd: mention what error caused a rename failure.  Remove \n's.
Rename AUDIT_TRIGGER_OPEN_NEW to AUDIT_TRIGGER_ROTATE_USER, and
add AUDIT_TRIGGER_ROTATE_KERNEL, to distinguish a rotate sourced
from the user and from the kernel.

Disable rate limiting on certain kernel-sourced triggers --
specifically rotate, because it won't be retransmitted, so if we
miss it, we won't see it again and the audit trail will grow
indefinitely.  This is bad.  Maintain rate limiting for other
space-related events.

Don't generate a syslog event for sigchld.  It's boring.
Likewise for a trigger read event, since we'll document the
details of the event in a separate log message.

Annotate a few bugs in how we handle mappings (we don't remove
old ones if deleted), and that we need to incrementally update
the kernel policy, not overwrite it, or we lose argc/argv/...
flags.  We need those to be stored in audit_control at some
point.

Fix a typo in the AUDITOFF_WARN string.

Update the maximum user record size to MAXAUDITDATA from 1k.

[git-p4: depot-paths = "//depot/projects/trustedbsd/openbsm/": change = 105886]