Reap children before entering main event loop
In the event we have thousands of audit trails to expuire, we go
through and expire each one, executing the audit_warn script. We
do not reap these children until we enter the main event loop.

This change adds a reap check right after the fork/exec operations,
if we have over 100 child processes that are waiting to be reaped,
do it. This hopefully fixes a denial of service condition when
starting up with many audit trail files.

In collaboration with:	asomers
Issue:	#35