Introduce au_set_eventdb_fd() interface
auditreduce runs inside a capsicum sandbox, which means certain types
of filtering will fail if they require file descriptors to the event
or class databases. Introduce au_set_eventdb_fd() which allows capsicum
consumers of libbsm to specify which file descriptor should be used for
the event database.

Teach auditreduce to use this function. Note we might scope it down and
grab the file descriptor only if -c is specified. But this is a humble
begining for now

Documentation to come.

Reported by: @ltning
Issue: https://github.com/openbsm/openbsm/issues/66